MAINT: merge into run.py due to similar purpose
