+++ a_package/simulation/__init__.py
@@ -1,10 +1,2 @@
 from .simulation import Simulation
 from .io import SimulationIO, NpyIO, Term
-from .setup import (
-    create_grid_from_config,
-    generate_surface_from_config,
-    build_capillary_args,
-    build_solver_args,
-    build_trajectory,
-    compute_liquid_volume,
-)
